backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.2
google-generativeai==0.3.2
groq==0.4.2
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
aiofiles==23.2.1
httpx==0.25.2
